Extract DiagnosisKeyBundler abstract class
